update gatsby docs and packages references to reflect breaking changes in emotion v11: - @emotion/core → @emotion/react - emotion-theming → moved into @emotion/react - babel-plugin-emotion → @emotion/babel-plugin - jest-emotion → @emotion/jest
update gatsby-plugin-emotion's autoLabel option from boolean to string
to reflect breaking changes in emotion v11.
