+++ packages/gatsby-plugin-emotion/src/gatsby-node.js
@@ -16,10 +16,10 @@ exports.pluginOptionsSchema = ({ Joi }) =>
       .description(
         `This option tells the plugin to inject source maps for use in browser dev tools in development.`
       ),
-    autoLabel: Joi.boolean()
-      .default(process.env.NODE_ENV !== `production`)
+    autoLabel: Joi.string()
+      .default(`dev-only`)
       .description(
-        `This option automatically adds the label property to styles so that class names generated by css or styled include the name of the variable the result is assigned to. You can read more about this option in babel-plugin-emotion's docs: https://emotion.sh/docs/babel-plugin-emotion#autolabel`
+        `This option automatically adds the label property to styles so that class names generated by css or styled include the name of the variable the result is assigned to. You can read more about this option in @emotion/babel-plugin's docs: https://emotion.sh/docs/@emotion/babel-plugin#autolabel`
       ),
     labelFormat: Joi.string()
       .default(`[local]`)
