+++ packages/gatsby-recipes/recipes/chakra-ui.mdx
@@ -8,9 +8,8 @@ This recipes sets up and configures [gatsby-plugin-chakra-ui](https://www.gatsby
 
 ## Installs necessary NPM packages
 
-<NPMPackage name="@emotion/core" />
+<NPMPackage name="@emotion/react" />
 <NPMPackage name="@emotion/styled" />
-<NPMPackage name="emotion-theming" />
 <NPMPackage name="@chakra-ui/core" />
 <NPMPackage name="gatsby-plugin-chakra-ui" />
