+++ packages/gatsby-recipes/recipes/emotion.mdx
@@ -11,7 +11,7 @@ This recipe:
 Installs necessary NPM packages.
 
 <NPMPackage name="gatsby-plugin-emotion" />
-<NPMPackage name="@emotion/core" />
+<NPMPackage name="@emotion/react" />
 <NPMPackage name="@emotion/styled" />
 
 ---
